arm64: dts: msm8939 SoC support
This commits adds support for Qualcomm MSM8939 SoC with
support for peripherals like I2C, eMMC, WLAN, MDSS and
so on.
The chip is largely based on MSM8916, allowing us to
repurpose many of the drivers.
Sadly, it is not capable of using PSCI, but it is possible
to bring up all eight cores with a different approach.