remoteproc: qcom_wcnss: Add performance corner regulator support
Processors such as the MSM8936, APQ8036, MSM8939 and APQ8039 have voltage
rails that are based on fixed performance levels as opposed to specific
voltage or current load settings.

These corner regulators are presented via the RPMD interface.

In the case of WCNSS chips that would be paired with the above SoCs vddmx
and vddcx are attached to these corner regulators.

This patch provides a binding to represent the corner regulators in a
similar way to how the PMIC vregs are represented.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>