arm64: dts: qcom: msm8939-samsung-a7: Add accelerometer/magnetometer
A7 uses a Bosch BMC150 accelerometer/magnetometer combo.
The chip provides two separate I2C devices for the accelerometer
and magnetometer that are already supported by the bmc150-accel
and bmc150-magn driver.

The difference is the way the sensor on A7 is mounted on the mainboard - set
the mount-matrix in the device-specific device tree part to handle that
difference.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>